daemon: introduce a new 'communication' module
The 'control' module grew to include both routines related to network
communication and ones related to RPC command, as the two error enums
present there outlined.

Split it into the existing 'control' and a new 'communication' mod.

This also largely reduces the size of the control.rs file by moving the
communication-related tests.
